% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{lookup}
\alias{lookup}
\alias{print.function}
\title{Lookup a function definiton}
\usage{
lookup(x, name = substitute(x), envir = environment(x) \%||\% baseenv(),
  all = FALSE, ...)
}
\arguments{
\item{x}{name or definition of a function}

\item{name}{function name if not given in x}

\item{envir}{the environment the function was defined in}

\item{all}{Whether to return just loaded definitions or all definitions}

\item{...}{Additional arguments passed to internal functions}
}
\description{
Lookup a function definiton
}

