% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{lookup_browse}
\alias{lookup_browse}
\title{Browse a lookup}
\usage{
lookup_browse(x = the$last_lookup)
}
\arguments{
\item{x}{The lookup to browse, by default will use the last url of the last lookup found.}
}
\description{
If a github or CRAN lookup, this will open a browser to the github code
mirror. If a local file will open the application associated with that file
type.
}

