% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{lookup_usage}
\alias{lookup_usage}
\title{Search for code usage on CRAN}
\usage{
lookup_usage(x)
}
\arguments{
\item{x}{The term to search}
}
\description{
This function uses GitHub's code search to search for a code usage in all
packages on CRAN. It will open a browser window with the results so they can
be explored.
}
\examples{
\dontrun{
lookup_usage("grep")
}
}

