% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{source_url}
\alias{source_url}
\title{Retrieve the source url(s) from a lookup object}
\usage{
source_url(x, ...)
}
\arguments{
\item{x}{The object to return the source_url from}

\item{...}{Additional arguments passed to methods}
}
\description{
Retrieve the source url(s) from a lookup object
}
